HSEARCH-1375 Simplify IndexFailureContextImpl and its builder
We're about to introduce more complexity with a type hierarchy,
so we better start from something simple.
